Add target support for cmake
Unlike some Stackoverflow answers [0] are suggesting, --target help is
not a decent source to pass as it will be missing the main targets even,
but will list all intermediate targets with no way to differentiate
useful targets.

Going straight for the zsh target completion source [1] shows that there
is a better way to get the targets, but it involves invoking ninja or
make.
Now the targets are labeled e.g. CXX_EXECUTABLE_LINKER_target, so we can
grep for executable to narrow down the list significantly.

[0] https://stackoverflow.com/a/30803359
[1] https://github.com/zsh-users/zsh-completions/blob/e61c9c14d6978191762e9586a0c882114e49221d/src/_cmake#L163