Populate quickfixlist with setqflist()
Both cexpr and cgetexpr use the global errorformat value rather than the
local value (like cfile does). This is a problem, because the ftplugin
files shipped with Vim set the compiler without "!", which means that
all following CompilerSet commands (such as setting the errorformat
option) will use the local scope rather than the global scope.

This means that just opening a Rust file for example and compiling will
not use the set errorformat value from the rust ftplugin setting the
cargo compiler, because it is invisible to cexpr only reading the global
value.

To work around this we could save the global errorformat value,
temporarily set it equal to the local value, then reset it again.
However, it is much cleaner to just use setqflist() which while it has
the same problem at least provides an optional "efm" argument to set the
errorformat value to be used.

Now we need to manually recreate the functionality of cexpr
automatically navigating to the first error, which also requires a check
for if the quickfix list is not empty.