Parse build.ninja for cmake target executables
This is a more reliable heuristic than just looking in the build
directory, even though parsing build.ninja is a bit cursed.

Unfortunately none of the ninja -t tools do print the TARGET_FILE as a
standalone, so manual parsing is necessary.

In the future for the default build we could also walk back the
dependency graph for the default -> all target, but for now the default
heuristic is still good enough