Fallback to ninja TARGET_FILE if target is unique
If we cannot find the binary with other heuristics, then we may still
have luck with TARGET_FILE.

We only try this fallback if the target is unique, because we cannot
reasonably advocate for iterating through random targets; that is just
calling for random code execution. With just a single target it's
still reasonable to find the associated binary so that it works more
often out of the box.