vimspector: Add support for string launch commands
In the future it might be better to actually parse the arguments, but
for now for most usecases ad-hoc conversion to array should be fine.
Without this it would use the first character of the string launch
command as program and the rest as arguments, which does not make any
sense.