Allow to configure skipping build phase for non-existing focus target
This implements the first part of the customization overhaul.
Overall, we want the user to be able to both:
- Keep the build phase and customize the launch target (this patch)
- Keep the launch target and customize the build phase

Similarly tangents can be drawn to the makeprg support.
With a2bdb051e63861d852a2b868f4f500f64455e315 implementing the necessary
library backend we can now use the makeprg from :compiler by using the
makeprg library backend and then customizing the target with
LaunchpadFocus.