Add git bisect integration
For now this is still in a pretty basic shape, as the popup is started
immediately, which can look hideous and obstruct other popups in some
cases.
In the future we can replace this with a callback from the launch
target.
Another possible improvement is going to be just forwarding the exit
code of the launch target.